requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
gunicorn==19.9.0
idna==2.8
pdf2image==1.6.0
Pillow==6.1.0
pytesseract==0.2.7
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
